Splitt ut oppsummeringvedtak til oppsummeringvedtak + oppsummeringvedtakinnhold
